+++ packages/fixie-common/tests/agent.test.ts
@@ -60,15 +60,19 @@ describe('FixieAgentBase Agent tests', () => {
           handle: 'fake-agent-handle-4',
         },
       ],
+      pageInfo: {
+        totalResultCount: 4,
+      },
     });
-    const agents = await FixieAgentBase.ListAgents({ client });
+    const result = await FixieAgentBase.ListAgents({ client });
     expect(mock.mock.calls[0][0].toString()).toStrictEqual(
       'https://fake.api.fixie.ai/api/v1/agents?offset=0&limit=100'
     );
-    expect(agents.length).toBe(4);
-    expect(agents[0].id).toBe('fake-agent-id-1');
-    expect(agents[0].handle).toBe('fake-agent-handle-1');
-    expect(agents[0].agentUrl()).toBe('https://console.fixie.ai/agents/fake-agent-id-1');
+    expect(result.total).toBe(4);
+    expect(result.agents.length).toBe(4);
+    expect(result.agents[0].id).toBe('fake-agent-id-1');
+    expect(result.agents[0].handle).toBe('fake-agent-handle-1');
+    expect(result.agents[0].agentUrl()).toBe('https://console.fixie.ai/agents/fake-agent-id-1');
   });
 
   it('ListAgents pagination works', async () => {
@@ -92,13 +96,17 @@ describe('FixieAgentBase Agent tests', () => {
           handle: 'fake-agent-handle-4',
         },
       ],
+      pageInfo: {
+        totalResultCount: 20,
+      },
     });
-    const agents = await FixieAgentBase.ListAgents({ client, offset: 0, limit: 4 });
+    const result = await FixieAgentBase.ListAgents({ client, offset: 0, limit: 4 });
     expect(mock.mock.calls[0][0].toString()).toStrictEqual('https://fake.api.fixie.ai/api/v1/agents?offset=0&limit=4');
-    expect(agents.length).toBe(4);
-    expect(agents[0].id).toBe('fake-agent-id-1');
-    expect(agents[0].handle).toBe('fake-agent-handle-1');
-    expect(agents[0].agentUrl()).toBe('https://console.fixie.ai/agents/fake-agent-id-1');
+    expect(result.total).toBe(20);
+    expect(result.agents.length).toBe(4);
+    expect(result.agents[0].id).toBe('fake-agent-id-1');
+    expect(result.agents[0].handle).toBe('fake-agent-handle-1');
+    expect(result.agents[0].agentUrl()).toBe('https://console.fixie.ai/agents/fake-agent-id-1');
   });
 
   it('CreateAgent works', async () => {