+++ packages/fixie/src/cli.ts
@@ -583,7 +583,7 @@ agent
     catchErrors(async (opts) => {
       const client = await AuthenticateOrLogIn({ apiUrl: program.opts().url });
       const result = await FixieAgent.ListAgents({ client, teamId: opts.teamId });
-      showResult(await Promise.all(result.map((agent) => agent.metadata)), program.opts().raw);
+      showResult(await Promise.all(result.agents.map((agent) => agent.metadata)), program.opts().raw);
     })
   );